requirements.doc.txt
sphinx
sphinx-gallery
sphinx_rtd_theme
sphinxcontrib-bibtex
numpydoc
matplotlib
pandas